requirements.txt
backend==0.2.4.1
cubemapencoder==0.0.0
dearpygui==1.11.1
imageio==2.35.1
kornia==0.7.3
matplotlib==3.7.5
mediapy==1.2.2
numpy==1.24.4
open3d==0.18.0
opencv_python==4.10.0.84
Pillow==11.1.0
plyfile==1.1
raytracing==0.1.0
scipy==1.15.1
setuptools==71.0.4
scikit-image==0.21.0
torch==2.0.0
torchvision==0.15.0
tqdm==4.66.2
trimesh==4.3.2
